Massive restructuring of Sprite loading system
Instead of specifying party, you now specify the sprite code of a mob,
which will be used to give it an array of images on creation. A script
can be used on Unix systems to generate the code for loading resources.
All mobs are currently set to dmw (Democratic, Male, White Congressman),
will be changed in the future.